+++ src/Libraries/SmartStore.Core/Extensions/ConversionExtensions.cs
@@ -13,6 +13,7 @@ using System.Runtime.Serialization.Formatters.Binary;
 
 using SmartStore.Utilities;
 using System.Diagnostics;
+using System.Security.Cryptography;
 
 namespace SmartStore
 {
@@ -540,6 +541,40 @@ namespace SmartStore
         public static string AsString(this byte[] bytes)
         {
             return Encoding.Default.GetString(bytes);
+        }
+
+              
+        /// <summary>
+        /// Computes the MD5 hash of a byte array
+        /// </summary>
+        /// <param name="value">The byte array to compute the hash for</param>
+        /// <returns>The hash value</returns>
+        //[DebuggerStepThrough]
+        public static string Hash(this byte[] value, bool toBase64 = false)
+        {
+            Guard.ArgumentNotNull(value, "value");
+
+            using (MD5 md5 = MD5.Create())
+            {
+
+                if (toBase64)
+                {
+                    byte[] hash = md5.ComputeHash(value);
+                    return System.Convert.ToBase64String(hash);
+                }
+                else
+                {
+                    StringBuilder sb = new StringBuilder();
+
+                    byte[] hashBytes = md5.ComputeHash(value);
+                    foreach (byte b in hashBytes)
+                    {
+                        sb.Append(b.ToString("x2").ToLower());
+                    }
+
+                    return sb.ToString();
+                }
+            }
         }
 
         #endregion